Move Integration annotation to test-common project
