+++ game-core/src/test/java/games/strategy/engine/chat/ChatIntegrationTest.java
@@ -12,6 +12,7 @@ import java.util.stream.IntStream;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.triplea.test.common.Integration;
 
 import games.strategy.engine.message.ChannelMessenger;
 import games.strategy.engine.message.IChannelMessenger;
@@ -25,7 +26,6 @@ import games.strategy.net.IServerMessenger;
 import games.strategy.net.MacFinder;
 import games.strategy.net.ServerMessenger;
 import games.strategy.sound.SoundPath;
-import games.strategy.test.Integration;
 
 @Integration
 public final class ChatIntegrationTest {
