+++ game-core/src/test/java/games/strategy/engine/lobby/server/db/AbstractControllerTestCase.java
@@ -1,7 +1,8 @@
 package games.strategy.engine.lobby.server.db;
 
+import org.triplea.test.common.Integration;
+
 import games.strategy.engine.config.lobby.TestLobbyPropertyReaders;
-import games.strategy.test.Integration;
 
 /**
  * Superclass for fixtures that test a DAO implementation.
