+++ game-core/src/test/java/games/strategy/net/MessengerIntegrationTest.java
@@ -16,8 +16,8 @@ import javax.annotation.concurrent.GuardedBy;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.triplea.test.common.Integration;
 
-import games.strategy.test.Integration;
 import games.strategy.util.Interruptibles;
 
 @Integration
